requirements.txt
dash==2.0.0
dash-bootstrap-components==0.13.1
folium==0.12.1
geojson==2.5.0
ipython-sql==0.4.0
jobspy==0.28.1
matplotlib==3.4.2
numpy==1.20.3
pandas==1.2.4
psycopg2-binary==2.9.1
waitress